+++ rpc/backend/call_tx.go
@@ -16,7 +16,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/core/vm"
+	corevm "github.com/ethereum/go-ethereum/core/vm"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -381,7 +381,7 @@ func (b *Backend) DoCall(
 	}
 
 	if res.Failed() {
-		if res.VmError != vm.ErrExecutionReverted.Error() {
+		if res.VmError != corevm.ErrExecutionReverted.Error() {
 			return nil, status.Error(codes.Internal, res.VmError)
 		}
 		return nil, evmtypes.NewExecErrorWithReason(res.Ret)