+++ x/erc20/keeper/msg_server.go
@@ -37,9 +37,7 @@ func (k Keeper) ConvertCoin(
 
 	// Remove token pair if contract is suicided
 	erc20 := common.HexToAddress(pair.Erc20Address)
-	acc := k.evmKeeper.GetAccountWithoutBalance(ctx, erc20)
-
-	if acc == nil || !acc.IsContract() {
+	if k.evmKeeper.IsEmptyAccount(ctx, erc20) {
 		k.DeleteTokenPair(ctx, pair)
 		k.Logger(ctx).Debug(
 			"deleting selfdestructed token pair from state",
@@ -79,9 +77,7 @@ func (k Keeper) ConvertERC20(
 
 	// Remove token pair if contract is suicided
 	erc20 := common.HexToAddress(pair.Erc20Address)
-	acc := k.evmKeeper.GetAccountWithoutBalance(ctx, erc20)
-
-	if acc == nil || !acc.IsContract() {
+	if k.evmKeeper.IsEmptyAccount(ctx, erc20) {
 		k.DeleteTokenPair(ctx, pair)
 		k.Logger(ctx).Debug(
 			"deleting selfdestructed token pair from state",