+++ x/evm/types/tracer.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/vm"
+	corevm "github.com/ethereum/go-ethereum/core/vm"
 	ethparams "github.com/ethereum/go-ethereum/params"
 )
 
@@ -22,7 +22,7 @@ const (
 
 // NewTracer creates a new Logger tracer to collect execution traces from an
 // EVM transaction.
-func NewTracer(tracer string, msg core.Message, cfg *ethparams.ChainConfig, height int64) vm.EVMLogger {
+func NewTracer(tracer string, msg core.Message, cfg *ethparams.ChainConfig, height int64) corevm.EVMLogger {
 	// TODO: enable additional log configuration
 	logCfg := &logger.Config{
 		Debug: true,
@@ -31,7 +31,7 @@ func NewTracer(tracer string, msg core.Message, cfg *ethparams.ChainConfig, heig
 	switch tracer {
 	case TracerAccessList:
 		const mergeNetsplit = true
-		preCompiles := vm.ActivePrecompiles(cfg.Rules(big.NewInt(height), mergeNetsplit))
+		preCompiles := corevm.ActivePrecompiles(cfg.Rules(big.NewInt(height), mergeNetsplit))
 		return logger.NewAccessListTracer(msg.AccessList(), msg.From(), *msg.To(), preCompiles)
 	case TracerJSON:
 		return logger.NewJSONLogger(logCfg, os.Stderr)
@@ -50,7 +50,7 @@ type TxTraceResult struct {
 	Error  string      `json:"error,omitempty"`  // Trace failure produced by the tracer
 }
 
-var _ vm.EVMLogger = &NoOpTracer{}
+var _ corevm.EVMLogger = &NoOpTracer{}
 
 // NoOpTracer is an empty implementation of vm.Tracer interface
 type NoOpTracer struct{}
@@ -63,7 +63,7 @@ func NewNoOpTracer() *NoOpTracer {
 // CaptureStart implements vm.Tracer interface
 //
 //nolint:revive // allow unused parameters to indicate expected signature
-func (dt NoOpTracer) CaptureStart(env *vm.EVM,
+func (dt NoOpTracer) CaptureStart(env *corevm.EVM,
 	from common.Address,
 	to common.Address,
 	create bool,
@@ -75,13 +75,13 @@ func (dt NoOpTracer) CaptureStart(env *vm.EVM,
 // CaptureState implements vm.Tracer interface
 //
 //nolint:revive // allow unused parameters to indicate expected signature
-func (dt NoOpTracer) CaptureState(pc uint64, op vm.OpCode, gas, cost uint64, scope *vm.ScopeContext, rData []byte, depth int, err error) {
+func (dt NoOpTracer) CaptureState(pc uint64, op corevm.OpCode, gas, cost uint64, scope *corevm.ScopeContext, rData []byte, depth int, err error) {
 }
 
 // CaptureFault implements vm.Tracer interface
 //
 //nolint:revive // allow unused parameters to indicate expected signature
-func (dt NoOpTracer) CaptureFault(pc uint64, op vm.OpCode, gas, cost uint64, scope *vm.ScopeContext, depth int, err error) {
+func (dt NoOpTracer) CaptureFault(pc uint64, op corevm.OpCode, gas, cost uint64, scope *corevm.ScopeContext, depth int, err error) {
 }
 
 // CaptureEnd implements vm.Tracer interface
@@ -92,7 +92,7 @@ func (dt NoOpTracer) CaptureEnd(output []byte, gasUsed uint64, tm time.Duration,
 // CaptureEnter implements vm.Tracer interface
 //
 //nolint:revive // allow unused parameters to indicate expected signature
-func (dt NoOpTracer) CaptureEnter(typ vm.OpCode, from common.Address, to common.Address, input []byte, gas uint64, value *big.Int) {
+func (dt NoOpTracer) CaptureEnter(typ corevm.OpCode, from common.Address, to common.Address, input []byte, gas uint64, value *big.Int) {
 }
 
 // CaptureExit implements vm.Tracer interface