+++ x/evm/vm/test_utils.go
@@ -1,0 +1,32 @@
+package vm
+
+import (
+	"github.com/EscanBE/evermint/v12/crypto/ethsecp256k1"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
+// NewAddrKey generates an Ethereum address and its corresponding private key.
+func NewAddrKey() (common.Address, *ethsecp256k1.PrivKey) {
+	privkey, _ := ethsecp256k1.GenerateKey()
+	key, err := privkey.ToECDSA()
+	if err != nil {
+		return common.Address{}, nil
+	}
+
+	addr := crypto.PubkeyToAddress(key.PublicKey)
+
+	return addr, privkey
+}
+
+// GenerateAddress generates an Ethereum address.
+func GenerateAddress() common.Address {
+	addr, _ := NewAddrKey()
+	return addr
+}
+
+// GenerateHash generates an Ethereum hash.
+func GenerateHash() common.Hash {
+	_, pk := NewAddrKey()
+	return common.BytesToHash(pk.Bytes())
+}